+++ HW4/core/src/edu/cs328/EntityManager.java
@@ -28,7 +28,7 @@ public class EntityManager extends EntitySystem {
     
     EntityManager(TextureAtlas atlas, World world) {
     	if (_instance != null) System.out.println("Creating multiple entity managers");
-    	_instance = this;
+        _instance = this;
     	this.atlas = atlas;
     	this.world = world;
     	for (int i = 0; i < 5; i++) {
@@ -64,7 +64,6 @@ public class EntityManager extends EntitySystem {
         	Box2dComponent b = b2dc.get(e);
             if(b.sprite != null)
                 b.draw(batch);
-
         }
     }
