+++ HW4/core/src/edu/cs328/SplashScreen.java
@@ -1,0 +1,43 @@
+package edu.cs328;
+
+import com.badlogic.gdx.Screen;
+
+/**
+ * Created by KnightPickles on 3/26/16.
+ */
+public class SplashScreen implements Screen {
+    @Override
+    public void show() {
+
+    }
+
+    @Override
+    public void render(float delta) {
+
+    }
+
+    @Override
+    public void resize(int width, int height) {
+
+    }
+
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void resume() {
+
+    }
+
+    @Override
+    public void hide() {
+
+    }
+
+    @Override
+    public void dispose() {
+
+    }
+}
